+++ CHANGELOG.rst
@@ -83,6 +83,7 @@ Fixed
     - CSS in project zone list (#1027)
     - Uncaught irodsbackend exceptions in ``TriggerZoneMoveTask`` (#1148)
     - Project list column retrieval failing with anonymous user (#1155)
+    - Inactive zones deleted from all projects on zone clear (#1150)
 - **Samplesheets**
     - MaxQuant results not correctly linked in ``pep_ms`` assay app (#1072)
     - Incorrect timeline event for ``sheet_create`` (#1080)